feat: 发送文件到服务器-2

+++ lib-clink/src/main/java/org/yangxin/socket/lib/box/StringReceivePacket.java
@@ -1,10 +1,6 @@
 package org.yangxin.socket.lib.box;
 
-import org.yangxin.socket.lib.core.ReceivePacket;
-
 import java.io.ByteArrayOutputStream;
-import java.io.Closeable;
-import java.io.IOException;
 
 /**
  * 字符串接收包
@@ -14,7 +10,7 @@ import java.io.IOException;
  */
 public class StringReceivePacket extends AbstractByteArrayReceivePacket<String> {
 
-    public StringReceivePacket(int length) {
+    public StringReceivePacket(long length) {
         super(length);
     }
